backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
numpy==1.26.4

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP Client
httpx==0.27.2
aiohttp==3.10.5

# Image Processing
Pillow==10.4.0
python-magic-bin==0.4.14  # for Windows

# Text-to-Speech (TTS)
google-cloud-texttospeech==2.14.1

# RAG (Retrieval-Augmented Generation)
chromadb==0.5.3
sentence-transformers==3.0.1
PyPDF2==3.0.1
torch>=2.0.0  # Required by sentence-transformers
transformers>=4.34.0  # Required by sentence-transformers
groq==0.4.2

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0